// src/lib/cores/alliancesCore/EAC/types.ts
export interface CustomsOffice {
  email: string
  website: string
  address: string
}
// ...
export interface MainInternationalSeaport {
  name: string
  website: string
  email: string
  twitter: string
  instagram: string
  linkedin: string
  apiEndpoint: string
  customsOffice: CustomsOffice
}